requirements.txt
google-generativeai==0.8.5
requests==2.32.3
python-dotenv>=1.0.0
flask>=2.3.0
pydantic==2.7.1
Pillow